Added part to receive account balance info

+++ src/main/java/com/steven/igtradingdemo/dto/PlaceTradeRequest.java
@@ -5,13 +5,14 @@ import com.iggroup.webapi.samples.client.rest.dto.positions.otc.createOTCPositio
 import lombok.Data;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @Data
 public class PlaceTradeRequest {
     @NotNull(message = "Market epic cannot be Null")
     private String marketEpic;
     @NotNull(message = "Position Size cannot be Null")
-    private Double positionSize;
+    private BigDecimal positionSize;
     @NotNull(message = "Trade Direction cannot be Null")
     private Direction tradeDirection;
 }
